+++ test/analysis_options_test.dart
@@ -25,7 +25,7 @@ void main() {
     expect(json.decode(content), <String, Object?>{});
   });
 
-  test('passthrough for some options', () {
+  test('passthrough for most options', () {
     final content = updatePassthroughOptions(
       original: '''
 analyzer:
@@ -44,7 +44,7 @@ formatter:
     );
     expect(json.decode(content), {
       'analyzer': {
-        'errors': {'uri_has_not_been_generated': 'ignore'},
+        'errors': {'todo': 'ignore', 'uri_has_not_been_generated': 'ignore'},
         'enable-experiment': ['macros'],
       },
       'formatter': {
@@ -55,25 +55,68 @@ formatter:
     });
   });
 
-  test('ignore include in original', () {
+  test('passthrough analysis errors except one', () {
+    final content = updatePassthroughOptions(
+      original: '''
+analyzer:
+  errors:
+    todo: ignore
+    another-todo: ignore
+    uri_has_not_been_generated: ignore
+''',
+      custom: '''
+linter:
+  rules:
+    - todo
+    - uri_has_not_been_generated
+''',
+    );
+    expect(json.decode(content), {
+      'analyzer': {
+        'errors': {
+          'another-todo': 'ignore',
+          'uri_has_not_been_generated': 'ignore',
+        },
+      },
+      'linter': isNotEmpty,
+    });
+  });
+
+  test('passthrough linter rules except one', () {
+    final content = updatePassthroughOptions(
+      original: '''
+linter:
+  rules:
+    avoid_relative_lib_imports: true
+    prefer_relative_imports: false
+    public_member_api_docs: false
+''',
+      custom: '''
+linter:
+  rules:
+    - prefer_relative_imports
+''',
+    );
+    expect(json.decode(content), {
+      'linter': {
+        'rules': {
+          'avoid_relative_lib_imports': true,
+          'prefer_relative_imports': true,
+          'public_member_api_docs': false,
+        },
+      },
+    });
+  });
+
+  test('update include from original', () {
     final content = updatePassthroughOptions(
       original: 'include: package:lints/other.yaml',
       custom: '',
     );
-    expect(json.decode(content), <String, Object?>{});
+    expect(json.decode(content), {'include': 'package:lints/other.yaml'});
   });
 
-  test('include only in custom: keepInclude=false', () {
-    final content = updatePassthroughOptions(
-      original: '',
-      custom: 'include: package:lints/other.yaml',
-    );
-    expect(json.decode(content), <String, Object?>{
-      'include': 'package:lints/other.yaml',
-    });
-  });
-
-  test('include only in custom: keepInclude=true', () {
+  test('include only in custom', () {
     final content = updatePassthroughOptions(
       original: '',
       custom: 'include: package:lints/other.yaml',
@@ -87,7 +130,6 @@ formatter:
     final content = updatePassthroughOptions(
       original: 'include: package:lints/other.yaml',
       custom: '',
-      keepInclude: true,
     );
     expect(json.decode(content), <String, Object?>{
       'include': 'package:lints/other.yaml',
@@ -98,7 +140,6 @@ formatter:
     final content = updatePassthroughOptions(
       original: 'include: package:lints/other.yaml',
       custom: 'include: package:lints/core.yaml',
-      keepInclude: true,
     );
     expect(json.decode(content), <String, Object?>{
       'include': 'package:lints/core.yaml',
@@ -106,11 +147,7 @@ formatter:
   });
 
   test('keep include without include value', () {
-    final content = updatePassthroughOptions(
-      original: '',
-      custom: '',
-      keepInclude: true,
-    );
+    final content = updatePassthroughOptions(original: '', custom: '');
     expect(json.decode(content), <String, Object?>{});
   });
 }