+++ test/license_test.dart
@@ -10,11 +10,7 @@ import 'package:test/test.dart';
 
 void main() {
   Future<void> expectFile(String path, List<String> expected) async {
-    final relativePath = path.substring('test/licenses/'.length);
-    final licenses = await detectLicenseInFile(
-      File(path),
-      relativePath: relativePath,
-    );
+    final licenses = await detectLicenseInFile(File(path));
     expect(licenses.map((e) => e.spdxIdentifier).toList(), expected);
   }
 
@@ -25,10 +21,7 @@ void main() {
   group('AGPL', () {
     test('explicit', () async {
       expect(
-        await detectLicenseInContent(
-          'GNU AFFERO GENERAL PUBLIC LICENSE',
-          relativePath: 'LICENSE',
-        ),
+        await detectLicenseInContent('GNU AFFERO GENERAL PUBLIC LICENSE'),
         isEmpty,
       );
       await expectFile('test/licenses/agpl_v3.txt', ['AGPL-3.0']);
@@ -40,7 +33,6 @@ void main() {
       expect(
         await detectLicenseInContent(
           '   Apache License\n     Version 2.0, January 2004\n',
-          relativePath: 'LICENSE',
         ),
         isEmpty,
       );
@@ -64,15 +56,11 @@ void main() {
       expect(
         await detectLicenseInContent(
           ['GNU GENERAL PUBLIC LICENSE', 'Version 2, June 1991'].join('\n'),
-          relativePath: 'LICENSE',
         ),
         isEmpty,
       );
       expect(
-        await detectLicenseInContent(
-          ['GNU GPL Version 2'].join('\n'),
-          relativePath: 'LICENSE',
-        ),
+        await detectLicenseInContent(['GNU GPL Version 2'].join('\n')),
         isEmpty,
       );
       await expectFile('test/licenses/gpl_v3.txt', ['GPL-3.0']);
@@ -84,7 +72,6 @@ void main() {
       expect(
         await detectLicenseInContent(
           '\nGNU LESSER GENERAL PUBLIC LICENSE\n    Version 3, 29 June 2007',
-          relativePath: 'LICENSE',
         ),
         isEmpty,
       );
@@ -95,19 +82,10 @@ void main() {
   group('MIT', () {
     test('explicit', () async {
       expect(
-        await detectLicenseInContent(
-          '\n\n   The MIT license\n\n blah...',
-          relativePath: 'LICENSE',
-        ),
+        await detectLicenseInContent('\n\n   The MIT license\n\n blah...'),
         isEmpty,
       );
-      expect(
-        await detectLicenseInContent(
-          'MIT license\n\n blah...',
-          relativePath: 'LICENSE',
-        ),
-        isEmpty,
-      );
+      expect(await detectLicenseInContent('MIT license\n\n blah...'), isEmpty);
 
       await expectFile('test/licenses/mit.txt', ['MIT']);
       await expectFile('test/licenses/mit_without_mit.txt', ['MIT']);
@@ -119,11 +97,22 @@ void main() {
       expect(
         await detectLicenseInContent(
           '\n\n   Mozilla Public License Version 2.0\n\n blah...',
-          relativePath: 'LICENSE',
         ),
         isEmpty,
       );
       await expectFile('test/licenses/mpl_v2.txt', ['MPL-2.0']);
+    });
+  });
+
+  group('Unicode v3', () {
+    test('explicit', () async {
+      expect(
+        await detectLicenseInContent(
+          '\n\n   UNICODE LICENSE V3\n\nCOPYRIGHT AND PERMISSION NOTICE\n',
+        ),
+        isEmpty,
+      );
+      await expectFile('test/licenses/unicode3-icu4x.txt', ['Unicode-3.0']);
     });
   });
 
@@ -132,7 +121,6 @@ void main() {
       expect(
         await detectLicenseInContent(
           '\n\n   This is free and unencumbered software released into the public domain.\n',
-          relativePath: 'LICENSE',
         ),
         isEmpty,
       );
@@ -142,10 +130,7 @@ void main() {
 
   group('unknown', () {
     test('empty content', () async {
-      expect(
-        await detectLicenseInContent('', relativePath: 'LICENSE'),
-        isEmpty,
-      );
+      expect(await detectLicenseInContent(''), isEmpty);
     });
   });